+++ src/lgl_1.cc
@@ -136,7 +136,7 @@ Status LGL_1::render()
     for (int i = 0; i < ncubes; ++i) {
         glm::mat4 model = glm::mat4(1.0f);
         model = glm::translate(model, cubePositions[i]);
-        float angle = 20.0f * i;
+        float angle = 20.0f * static_cast<float>(i);
         model = glm::rotate(model, glm::radians(angle), glm::vec3(1.0f, 0.3f, 0.5f));
 
         m_texturedShader.set_mat4("model", model);
